try to explicitly convert from long long int to int

+++ src/bigmemory.cpp
@@ -1413,7 +1413,7 @@ SEXP CGetType(SEXP bigMatAddr)
 {
   Rcpp::XPtr<BigMatrix> pMat(bigMatAddr);
   int ret = pMat->matrix_type();
-  return Rcpp::wrap(ret);
+  return Rcpp::wrap((int)ret);
 }
 
 // not currently used?!?!?!
